+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/request/handlers/UtxoPersistFilteredTransactionRequestHandler.kt
@@ -7,7 +7,7 @@ import net.corda.data.ledger.persistence.PersistFilteredTransactionsAndSignature
 import net.corda.data.persistence.EntityResponse
 import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
 import net.corda.ledger.common.data.transaction.filtered.FilteredTransaction
-import net.corda.ledger.libs.utxo.UtxoPersistenceService
+import net.corda.ledger.libs.persistence.utxo.UtxoPersistenceService
 import net.corda.ledger.persistence.common.LedgerPersistenceUtils.findAccount
 import net.corda.ledger.persistence.common.RequestHandler
 import net.corda.messaging.api.records.Record
