+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/utxo/impl/request/handlers/UtxoPersistSignedGroupParametersIfDoNotExistRequestHandler.kt
@@ -5,10 +5,9 @@ import net.corda.data.flow.event.external.ExternalEventContext
 import net.corda.data.ledger.persistence.PersistSignedGroupParametersIfDoNotExist
 import net.corda.data.persistence.EntityResponse
 import net.corda.flow.external.events.responses.factory.ExternalEventResponseFactory
-import net.corda.ledger.libs.utxo.SignatureSpec
-import net.corda.ledger.libs.utxo.SignatureWithKey
-import net.corda.ledger.libs.utxo.SignedGroupParameters
-import net.corda.ledger.libs.utxo.UtxoPersistenceService
+import net.corda.ledger.libs.persistence.utxo.SignatureWithKey
+import net.corda.ledger.libs.persistence.utxo.SignedGroupParameters
+import net.corda.ledger.libs.persistence.utxo.UtxoPersistenceService
 import net.corda.ledger.persistence.common.RequestHandler
 import net.corda.messaging.api.records.Record
 
@@ -34,7 +33,7 @@ class UtxoPersistSignedGroupParametersIfDoNotExistRequestHandler(
                     signedGroupParameters.mgmSignature.publicKey.array(),
                     signedGroupParameters.mgmSignature.bytes.array()
                 ),
-                SignatureSpec(
+                net.corda.ledger.libs.persistence.utxo.SignatureSpec(
                     signedGroupParameters.mgmSignatureSpec.signatureName,
                     signedGroupParameters.mgmSignatureSpec.customDigestName,
                     signedGroupParameters.mgmSignatureSpec.params.bytes.array()
