+++ components/ledger/ledger-persistence/src/test/kotlin/net/corda/ledger/persistence/consensual/ConsensualComponentGroupMapperTest.kt
@@ -1,6 +1,6 @@
 package net.corda.ledger.persistence.consensual
 
-import net.corda.ledger.libs.common.mapToComponentGroups
+import net.corda.ledger.libs.persistence.common.mapToComponentGroups
 import net.corda.ledger.persistence.consensual.impl.ConsensualComponentGroupMapper
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertEquals
