+++ components/ledger/ledger-persistence/src/test/kotlin/net/corda/ledger/persistence/utxo/UtxoComponentGroupMapperTest.kt
@@ -1,7 +1,7 @@
 package net.corda.ledger.persistence.utxo
 
-import net.corda.ledger.libs.common.mapToComponentGroups
-import net.corda.ledger.libs.utxo.impl.UtxoComponentGroupMapper
+import net.corda.ledger.libs.persistence.common.mapToComponentGroups
+import net.corda.ledger.libs.persistence.utxo.impl.UtxoComponentGroupMapper
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
