+++ components/ledger/ledger-persistence/src/test/kotlin/net/corda/ledger/persistence/utxo/impl/UtxoPersistenceServiceImplTest.kt
@@ -8,11 +8,11 @@ import net.corda.crypto.testkit.SecureHashUtils.randomSecureHash
 import net.corda.ledger.common.data.transaction.PrivacySalt
 import net.corda.ledger.common.data.transaction.TransactionMetadataImpl
 import net.corda.ledger.common.data.transaction.TransactionStatus
-import net.corda.ledger.libs.json.DefaultContractStateVaultJsonFactory
-import net.corda.ledger.libs.utxo.CustomRepresentation
-import net.corda.ledger.libs.utxo.UtxoRepository
-import net.corda.ledger.libs.utxo.UtxoTransactionReader
-import net.corda.ledger.libs.utxo.impl.UtxoPersistenceServiceImpl
+import net.corda.ledger.libs.persistence.json.DefaultContractStateVaultJsonFactory
+import net.corda.ledger.libs.persistence.utxo.CustomRepresentation
+import net.corda.ledger.libs.persistence.utxo.UtxoRepository
+import net.corda.ledger.libs.persistence.utxo.UtxoTransactionReader
+import net.corda.ledger.libs.persistence.utxo.impl.UtxoPersistenceServiceImpl
 import net.corda.ledger.persistence.json.impl.ContractStateVaultJsonFactoryRegistryImpl
 import net.corda.ledger.persistence.json.impl.DefaultContractStateVaultJsonFactoryImpl
 import net.corda.utilities.time.UTCClock
