+++ libs/persistence/src/main/java/net/corda/ledger/libs/persistence/common/package-info.java
@@ -1,6 +1,6 @@
 @Export
 @QuasarIgnoreAllPackages
-package net.corda.ledger.libs.utxo;
+package net.corda.ledger.libs.persistence.common;
 
 import co.paralleluniverse.quasar.annotations.QuasarIgnoreAllPackages;
 import org.osgi.annotation.bundle.Export;
