+++ libs/persistence/src/main/java/net/corda/ledger/libs/persistence/json/package-info.java
@@ -1,6 +1,6 @@
 @Export
 @QuasarIgnoreAllPackages
-package net.corda.ledger.libs.utxo.impl;
+package net.corda.ledger.libs.persistence.json;
 
 import co.paralleluniverse.quasar.annotations.QuasarIgnoreAllPackages;
 import org.osgi.annotation.bundle.Export;
