+++ libs/persistence/src/main/java/net/corda/ledger/libs/persistence/utxo/impl/package-info.java
@@ -1,6 +1,6 @@
 @Export
 @QuasarIgnoreAllPackages
-package net.corda.ledger.libs.json;
+package net.corda.ledger.libs.persistence.utxo.impl;
 
 import co.paralleluniverse.quasar.annotations.QuasarIgnoreAllPackages;
 import org.osgi.annotation.bundle.Export;
