+++ libs/persistence/src/main/kotlin/net/corda/ledger/libs/persistence/common/InconsistentLedgerStateException.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.libs.common
+package net.corda.ledger.libs.persistence.common
 
 import net.corda.v5.base.exceptions.CordaRuntimeException
