+++ libs/persistence/src/main/kotlin/net/corda/ledger/libs/persistence/utxo/CustomRepresentation.kt
@@ -1,3 +1,3 @@
-package net.corda.ledger.libs.utxo
+package net.corda.ledger.libs.persistence.utxo
 
 data class CustomRepresentation(val json: String)
