+++ libs/persistence/src/main/kotlin/net/corda/ledger/libs/persistence/utxo/SignatureSpec.kt
@@ -1,11 +1,11 @@
-package net.corda.ledger.libs.utxo
+package net.corda.ledger.libs.persistence.utxo
 
 data class SignatureSpec(val signatureName: String, val customDigestName: String?, val params: ByteArray?) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
 
-        other as SignatureSpec
+        other as net.corda.ledger.libs.persistence.utxo.SignatureSpec
 
         if (signatureName != other.signatureName) return false
         if (customDigestName != other.customDigestName) return false
