+++ libs/persistence/src/main/kotlin/net/corda/ledger/libs/persistence/utxo/UtxoTransactionReader.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.libs.utxo
+package net.corda.ledger.libs.persistence.utxo
 
 import net.corda.ledger.common.data.transaction.PrivacySalt
 import net.corda.ledger.common.data.transaction.TransactionMetadataInternal
@@ -32,7 +32,7 @@ interface UtxoTransactionReader {
 
     fun getVisibleStates(): Map<Int, StateAndRef<ContractState>>
 
-    fun getConsumedStates(persistenceService: UtxoPersistenceService): List<StateAndRef<ContractState>>
+    fun getConsumedStates(persistenceService: net.corda.ledger.libs.persistence.utxo.UtxoPersistenceService): List<StateAndRef<ContractState>>
 
     fun getConsumedStateRefs(): List<StateRef>
