+++ libs/persistence/src/main/kotlin/net/corda/ledger/libs/persistence/utxo/impl/PostgresUtxoQueryProvider.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.libs.utxo.impl
+package net.corda.ledger.libs.persistence.utxo.impl
 
 import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup
