+++ testing/ledger/ledger-hsqldb/src/main/kotlin/net/corda/testing/ledger/utxo/HsqldbUtxoQueryProvider.kt
@@ -1,7 +1,7 @@
 package net.corda.testing.ledger.utxo
 
-import net.corda.ledger.libs.utxo.impl.AbstractUtxoQueryProvider
-import net.corda.ledger.libs.utxo.impl.UtxoQueryProvider
+import net.corda.ledger.libs.persistence.utxo.impl.AbstractUtxoQueryProvider
+import net.corda.ledger.libs.persistence.utxo.impl.UtxoQueryProvider
 import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup
 import net.corda.orm.DatabaseTypeProvider
 import net.corda.orm.DatabaseTypeProvider.Companion.HSQLDB_TYPE_FILTER
